IoTServicesCode/microservice/measurements_microservice/requirements.txt
mysql-connector-python
Flask
Flask-Cors
PyYAML